+++ src/adplug/rad.cpp
@@ -166,5 +166,5 @@ size_t CradLoader::framesUntilUpdate() const {
     if (currentTempo() != 0)
         return SampleRate / currentTempo();
     else
-        return SampleRate / 18.2;
+        return static_cast<size_t>(SampleRate / 18.2);
 }
